moved print conll to own script
